feat: implement edit-event features and routes
