+++ src/components/edit-event/edit-event-booking-options.tsx
@@ -1,15 +1,15 @@
 import { useEventData } from "@/hooks/useEventData/useEventData";
 import React from "react";
-import { Button } from "./ui/button";
+import { Button } from "../ui/button";
 import { EditEventWindow } from "@/hooks/useEventData/event-data-context";
 import { ChevronLeft } from "lucide-react";
-import { Form } from "./ui/form";
+import { Form } from "../ui/form";
 import { z } from "zod";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 
 const EditEventBookingOptions = () => {
-  const { eventBookingOptions, changeWindow, currentWindow } = useEventData();
+  const { changeWindow, currentWindow } = useEventData();
 
   const formSchema = z.object({});
